reduce complexity - fantasy land
