+++ composer.json
@@ -20,8 +20,7 @@
     "psr-0": {
       "Common": "src",
       "Monad": "src",
-      "Functor": "src",
-      "Applicative": "src"
+      "FantasyLand": "src"
     },
     "files": [
       "src/Functional/functions.php",
