+++ example/ApplicativeFunctorTest.php
@@ -1,14 +1,14 @@
 <?php
 namespace example;
 
-use Applicative;
+use Monad;
 use Functional as f;
 
 class ApplicativeFunctorTest extends \PHPUnit_Framework_TestCase
 {
     public function test_it_should_apply_every_function_in_collection_with_every_item_in_second()
     {
-        $collectionA = Applicative\Collection::create([
+        $collectionA = Monad\Collection::create([
             function ($a) {
                 return 3 + $a;
             },
@@ -16,14 +16,14 @@ class ApplicativeFunctorTest extends \PHPUnit_Framework_TestCase
                 return 4 + $a;
             },
         ]);
-        $collectionB = Applicative\Collection::create([
+        $collectionB = Monad\Collection::create([
             1,
             2
         ]);
 
         $result = $collectionA->ap($collectionB);
 
-        $this->assertInstanceOf(Applicative\Collection::class, $result);
+        $this->assertInstanceOf(Monad\Collection::class, $result);
         $this->assertEquals([4, 5, 5, 6], f\valueOf($result));
     }
 }
