+++ example/FunctorCollectionTest.php
@@ -1,14 +1,14 @@
 <?php
 namespace example;
 
-use Functor;
+use Monad;
 use Functional as f;
 
-class FunctorCollectionTest extends \PHPUnit_Framework_TestCase
+class MonadCollectionTest extends \PHPUnit_Framework_TestCase
 {
     public function test_it_should_return_new_map()
     {
-        $collection = Functor\Collection::create([
+        $collection = Monad\Collection::create([
             ['id' => 1, 'name' => 'One'],
             ['id' => 2, 'name' => 'Two'],
             ['id' => 3, 'name' => 'Three'],
