+++ spec/Monad/IdentitySpec.php
@@ -13,7 +13,7 @@ class IdentitySpec extends ObjectBehavior
     {
         $this->beConstructedWith(null);
         $this->shouldHaveType('Monad\Identity');
-        $this->shouldHaveType('Monad\MonadInterface');
+        $this->shouldHaveType('FantasyLand\MonadInterface');
     }
 
     public function it_should_bind_value_from_constructor_to_given_function()
@@ -62,5 +62,69 @@ class IdentitySpec extends ObjectBehavior
         });
 
         $right->bind('Functional\identity')->shouldReturn($left->bind('Functional\identity'));
+    }
+
+    public function it_should_obey_identity_law_applicative()
+    {
+        $this->beConstructedWith(function ($x) {
+            return $x;
+        });
+        $result = $this->ap($this::create(1));
+
+        $result->valueOf()->shouldReturn(1);
+    }
+
+    public function it_should_obey_homomorphism_law_applicative()
+    {
+        $id = function ($x) {
+            return $x;
+        };
+        $this->beConstructedWith($id);
+        $result = $this->ap($this::create(1));
+
+        $result->valueOf()->shouldReturn(
+            $this::create($id(1))->valueOf()
+        );
+    }
+
+    public function it_should_obey_interchange_law_applicative()
+    {
+        $y = 1;
+        $f = function ($x) {
+            return $x / 2;
+        };
+
+        $this->beConstructedWith($f);
+        $result = $this->ap($this::create($y));
+
+        $result->valueOf()->shouldReturn(
+            $this::create(function ($f) use ($y) {
+                return $f($y);
+            })->ap($this)->valueOf()
+        );
+    }
+
+    public function it_should_obey_identity_law_functor()
+    {
+        $this->beConstructedWith(1);
+        $result = $this->map(function ($x) {
+            return $x;
+        });
+
+        $result->valueOf()->shouldReturn(1);
+    }
+
+    public function it_should_obey_composition_law_functor()
+    {
+        $a = function ($x) {
+            return $x + 1;
+        };
+        $b = function ($x) {
+            return $x + 2;
+        };
+        $this->beConstructedWith(1);
+
+        $result = $this->map($a)->map($b);
+        $result->valueOf()->shouldReturn($b($a(1)));
     }
 }
