--- src/Functor/Identity.php
@@ -1,13 +1,0 @@
-<?php
-namespace Functor;
-
-use Common;
-
-class Identity implements FunctorInterface, Common\ValueOfInterface
-{
-    use Common\CreateTrait;
-    use Common\ValueOfTrait;
-    use MapTrait;
-
-    const create = 'Functor\Identity::create';
-}
