--- src/Monad/Feature/LiftInterface.php
@@ -1,15 +1,0 @@
-<?php
-namespace Monad\Feature;
-
-use Monad\MonadInterface;
-
-interface LiftInterface
-{
-    /**
-     * Converts values returned by regular function to monadic value.
-     *
-     * @param callable $transformation
-     * @return MonadInterface
-     */
-    public function lift(callable $transformation);
-}
