+++ src/Monad/Maybe/Maybe.php
@@ -4,13 +4,12 @@ namespace Monad\Maybe;
 use Monad;
 use Common;
 use Functor;
+use FantasyLand;
 use Applicative;
 
 interface Maybe extends
-    Monad\MonadInterface,
-    Common\ValueOfInterface,
-    Functor\FunctorInterface,
-    Applicative\ApplicativeInterface
+    FantasyLand\MonadInterface,
+    Common\ValueOfInterface
 {
     /**
      * Handle situation when error occur in monad computation chain.
